backend/app/core/config.py
import os
from dotenv import load_dotenv

# Carga las variables desde el archivo .env
load_dotenv()

# Define y exporta cada variable
DATABASE_URL = os.getenv("DATABASE_URL")
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES"))