backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart